.013
I captured the ratingvalue from the itunes site as a float and added it to the podcast_df dataframe. I also cleaned up some of the code on the it_getdata.py file to make it tabbed rather than spaces since py program had a bunch of errors due to it having spaces.